+++ src/Perplex.ContentBlocks.Core/Definitions/Api/ApiContentBlockDefinition.cs
@@ -4,6 +4,7 @@ public class ApiContentBlockDefinition
     public Guid Id { get; set; }
     public string Name { get; set; } = "";
     public string Description { get; set; } = "";
+    public string BlockNameTemplate { get; set; } = "";
     public string PreviewImage { get; set; } = "";
     public Guid ElementTypeKey { get; set; }
     public IEnumerable<Guid> CategoryIds { get; set; } = [];