+++ src/Perplex.ContentBlocks.Core/Definitions/ContentBlockDefinition.cs
@@ -5,6 +5,7 @@ public class ContentBlockDefinition : IContentBlockDefinition
     public Guid Id { get; set; }
     public string Name { get; set; } = "";
     public string Description { get; set; } = "";
+    public string BlockNameTemplate { get; } = "";
     public string PreviewImage { get; set; } = "";
     public Guid ElementTypeKey { get; set; }
     public IEnumerable<Guid> CategoryIds { get; set; } = [];